Rename behavior file
