+++ dds/src/dds/domain/dcps_service.rs
@@ -851,7 +851,7 @@ fn on_offered_incompatible_qos_communication_change(data_writer_node: DataWriter
         THE_DDS_DOMAIN_PARTICIPANT_FACTORY.get_participant_mut(
             &data_writer_node.parent_participant().prefix(),
             |dp| {
-                crate::implementation::dds_impl::node_user_defined_data_writer::get_offered_incompatible_qos_status(
+                crate::implementation::dds_impl::behavior_user_defined_data_writer::get_offered_incompatible_qos_status(
                     dp.ok_or(DdsError::AlreadyDeleted)?,
                     data_writer_node.guid(),
                     data_writer_node.parent_publisher(),
@@ -927,7 +927,7 @@ fn on_publication_matched_communication_change(data_writer_node: DataWriterNode)
         THE_DDS_DOMAIN_PARTICIPANT_FACTORY.get_participant_mut(
             &data_writer_node.parent_participant().prefix(),
             |dp| {
-                crate::implementation::dds_impl::node_user_defined_data_writer::get_publication_matched_status(dp.ok_or(DdsError::AlreadyDeleted)?,data_writer_node.guid(),
+                crate::implementation::dds_impl::behavior_user_defined_data_writer::get_publication_matched_status(dp.ok_or(DdsError::AlreadyDeleted)?,data_writer_node.guid(),
                 data_writer_node.parent_publisher(),)
             },
         )
